+++ CotEditor/Sources/CELayoutManager.m
@@ -158,6 +158,7 @@ static BOOL usesTextFontForInvisibles;
         // フォントサイズは随時変更されるため、表示時に取得する
         CTFontRef font = (__bridge CTFontRef)[self textFont];
         NSColor *color = [[self theme] invisiblesColor];
+        CGFloat baselineOffset = [self defaultBaselineOffsetForFont:[self textFont]];
         
         // for other invisibles
         NSFont *replacementFont;  // delay creating font till it's really needed
@@ -256,7 +257,7 @@ static BOOL usesTextFontForInvisibles;
             }
             
             // add invisible char path
-            NSPoint point = [self pointToDrawGlyphAtIndex:glyphIndex];
+            NSPoint point = [self pointToDrawGlyphAtIndex:glyphIndex verticalOffset:baselineOffset];
             CGAffineTransform translate = CGAffineTransformMakeTranslation(point.x, -point.y);
             CGPathAddPath(paths, &translate, glyphPath);
         }
@@ -506,17 +507,18 @@ static BOOL usesTextFontForInvisibles;
 
 //------------------------------------------------------
 /// グリフを描画する位置を返す
-- (NSPoint)pointToDrawGlyphAtIndex:(NSUInteger)glyphIndex
+- (NSPoint)pointToDrawGlyphAtIndex:(NSUInteger)glyphIndex verticalOffset:(CGFloat)offset
 //------------------------------------------------------
 {
-    NSPoint drawPoint = [self locationForGlyphAtIndex:glyphIndex];
-    NSPoint lineOrigin = [self lineFragmentRectForGlyphAtIndex:glyphIndex
-                                                effectiveRange:NULL
-                                       withoutAdditionalLayout:YES].origin;
+    NSPoint origin = [self lineFragmentRectForGlyphAtIndex:glyphIndex
+                                            effectiveRange:NULL
+                                   withoutAdditionalLayout:YES].origin;
+    NSPoint glyphLocation = [self locationForGlyphAtIndex:glyphIndex];
     
-    drawPoint.y += lineOrigin.y;
+    origin.x += glyphLocation.x;
+    origin.y += offset;
     
-    return drawPoint;
+    return origin;
 }