+++ CHANGELOG.md
@@ -12,6 +12,7 @@ develop
 
 ### Fixes
 
+- Fix an issue where the baseline of new line invisible characters was wrong if line is empty.
 - Fix an issue where syntax highlighted control character was sometime not colored in the invisible color.
 - [beta] Address syntax highlighiting issue with multiple lines.
 
@@ -51,7 +52,6 @@ develop
 - Address syntax highlighiting issue with multiple lines.
 - Fix an issue where text view drawing was distorted while resizing window.
 - Fix an issue where application could crash on window restoration.
-- Fix an issue where the baseline of new line invisible characters was wrong if line is empty.
 - Fix some typos in syntax styles “Julia” and “SQL”.